Revert "Revert "[Hotfix] sitemap route""
