+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "jiho",
-  "version": "1.9.2",
+  "version": "1.9.3",
   "private": true,
   "dependencies": {
     "@babel/preset-env": "^7.24.4",
@@ -34,6 +34,7 @@
     "start:prod": "export PORT=3005 && npm run build && serve -s build",
     "start:prod:ssr": "npm run build && npm run build:ssr",
     "build": "GENERATE_SOURCEMAP=false react-scripts build",
+    "prebuild": "node server/sitemap",
     "wbuild": "set\"GENERATE_SOURCEMAP=false\" && react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
