+++ public/sitemap.xml
@@ -1,0 +1,289 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://www.sitemaps.org/schemas/sitemap/0.9 http://www.sitemaps.org/schemas/sitemap/0.9/sitemap.xsd"
+    xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+   <url>
+    <loc>https://uosjudo.com/photo</loc>
+		<changefreq>weekly</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/31</loc>
+		<lastmod>2022-09-14</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/32</loc>
+		<lastmod>2022-09-16</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/33</loc>
+		<lastmod>2022-09-19</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/34</loc>
+		<lastmod>2022-09-21</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/35</loc>
+		<lastmod>2022-09-23</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/36</loc>
+		<lastmod>2022-09-26</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/37</loc>
+		<lastmod>2022-09-28</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/38</loc>
+		<lastmod>2022-09-30</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/39</loc>
+		<lastmod>2022-10-03</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/40</loc>
+		<lastmod>2022-10-05</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/41</loc>
+		<lastmod>2022-10-08</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/42</loc>
+		<lastmod>2022-11-02</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/43</loc>
+		<lastmod>2022-11-08</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/44</loc>
+		<lastmod>2022-11-09</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/45</loc>
+		<lastmod>2022-11-11</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/46</loc>
+		<lastmod>2022-11-12</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/47</loc>
+		<lastmod>2022-11-14</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/48</loc>
+		<lastmod>2022-11-16</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/49</loc>
+		<lastmod>2022-11-18</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/50</loc>
+		<lastmod>2022-11-23</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/51</loc>
+		<lastmod>2022-11-28</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/52</loc>
+		<lastmod>2022-11-30</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/photo/53</loc>
+		<lastmod>2022-12-02</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news</loc>
+		<changefreq>yearly</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022</loc>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/54</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/1</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/3</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/2</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/8</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/26</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/27</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/24</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/5</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/19</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/29</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/28</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/25</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/6</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/23</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/7</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/22</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/20</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/63</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/4</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2022/21</loc>
+		<lastmod>2022-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2023</loc>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2023/64</loc>
+		<lastmod>2023-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2023/65</loc>
+		<lastmod>2023-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2023/68</loc>
+		<lastmod>2023-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2023/69</loc>
+		<lastmod>2023-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2023/70</loc>
+		<lastmod>2023-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2023/66</loc>
+		<lastmod>2023-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2023/71</loc>
+		<lastmod>2023-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/news/2023/67</loc>
+		<lastmod>2023-03-01</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/notice</loc>
+  </url>
+ <url>
+    <loc>https://uosjudo.com/notice/61</loc>
+		<lastmod>2023-05-07</lastmod>
+		<changefreq>never</changefreq>
+  </url>
+</urlset>
