+++ server/index.js
@@ -4,5 +4,4 @@ require("@babel/register")({
   presets: ["@babel/preset-env", "@babel/preset-react"],
 });
 
-require("./sitemap");
 require("./server.js");
